+++ src/autocompletion/VariableFacet.js
@@ -1,0 +1,6 @@
+import { Facet } from '@codemirror/state';
+
+/**
+ * @type {Facet<import('..').Variable[]>} Variable
+ */
+export const variablesFacet = Facet.define();